Updates xdrfile.h, xdrfile_trr.h, and xdrfile_trr.c to double precision

+++ fortran/lib/xdrfile-1.1.4-d/include/xdrfile.h
@@ -105,8 +105,8 @@ extern "C"
 	extern char *exdr_message[exdrNR];
 
 #define DIM 3
-	typedef float matrix[DIM][DIM];
-	typedef float rvec[DIM];
+	typedef double matrix[DIM][DIM];
+	typedef double rvec[DIM];
 	typedef int   mybool;
